+++ users/views.py
@@ -3,6 +3,9 @@ from django.urls import reverse_lazy
 from users.forms import RegistrationForm
 from users.models import EmailUser
 from django.contrib.auth.views import LoginView
+from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
+from django.views.generic import TemplateView
 
 class RegisterView(CreateView):
     model = EmailUser
@@ -12,3 +15,7 @@ class RegisterView(CreateView):
 
 class UserLoginView(LoginView):
     template_name = "users/login.html"
+
+@method_decorator(login_required, name='dispatch')
+class DashboardView(TemplateView):
+    template_name = "users/dashboard.html"